+++ libdariadb/storage/index.h
@@ -20,7 +20,7 @@ struct IndexHeader {
   bool is_closed;
   uint64_t id_bloom;   // bloom filter of Meas.id
   uint64_t flag_bloom; // bloom filter of Meas.flag
-  uint32_t transaction; 
+  uint64_t transaction;
 };
 
 struct IndexReccord {
@@ -33,7 +33,7 @@ struct IndexReccord {
   uint64_t offset; // offset in bytes of chunk in page
   bool is_init;    // is init :)
   bool  is_transaction_closed : 1;
-  uint32_t transaction; 
+  uint64_t transaction;
   bool commit : 1;
 };
 #pragma pack(pop)
