+++ libdariadb/storage/page.h
@@ -21,7 +21,7 @@ struct PageHeader {
   dariadb::Time maxTime;
   uint64_t max_chunk_id; // max(chunk->id)
 
-  uint32_t transaction;
+  uint64_t transaction;
   bool _under_transaction;
 };
 #pragma pack(pop)
@@ -58,9 +58,9 @@ public:
   virtual append_result append(const Meas &value) override;
   virtual void flush() override;
 
-  void begin_transaction(uint32_t num);
-  void commit_transaction(uint32_t num);
-  void rollback_transaction(uint32_t num);
+  void begin_transaction(uint64_t num);
+  void commit_transaction(uint64_t num);
+  void rollback_transaction(uint64_t num);
 private:
   void flush_current_chunk();
   void init_chunk_index_rec(Chunk_Ptr ch);
