+++ libdariadb/storage/page_manager.h
@@ -64,9 +64,9 @@ public:
 
   void fsck(bool force_check = true); //if false - check files openned for write-only
 
-  uint32_t begin_transaction();
-  void commit_transaction(uint32_t num);
-  void rollback_transaction(uint32_t num);
+  uint64_t begin_transaction();
+  void commit_transaction(uint64_t num);
+  void rollback_transaction(uint64_t num);
 private:
   static PageManager *_instance;
   class Private;
